+++ src/mongodb/documents/ingredient.document.ts
@@ -1,6 +1,0 @@
-import { Document } from 'mongoose';
-
-export interface IngredientDocument extends Document {
-    readonly name: string;
-    readonly category: string; // TODO: Category type
-}