+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.25.0] - 2024-02-15
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): Ingredient as a string instead of object.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): Ingredient schema, repository and DTO.
+
 ## [1.24.1] - 2024-02-15
 ### Fixed
 - [Marek Kurańda](https://github.com/mjkuranda): `/users/refreshTokens` endpoint sending no content.